update template & add disabled/activate
